python/paragon_mnist_service/requirements.txt
fastapi
uvicorn
pillow
paragon_py
